dev to main

+++ src/Controller/AdminMemberController.php
@@ -167,21 +167,23 @@ class AdminMemberController extends AbstractAdminController
         }
     }
 
-    public function delete(int $id): void
+    public function delete(): void
     {
-        // Check Post Resquest
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            // delete en bdd
-            $memberManager = new MemberManager();
-            $member = $memberManager->selectOneById($id);
+            if (!empty($_POST['id'])) {
+                $id = $_POST['id'];
+                // delete en bdd
+                $memberManager = new MemberManager();
+                $member = $memberManager->selectOneById($id);
 
-            // supprimer un fichier existant
-            $this->deleteFile($member['image']);
-            // delete row
-            $memberManager->delete($id);
+                // supprimer un fichier existant
+                $this->deleteFile($member['image']);
+                // delete row
+                $memberManager->delete($id);
+            }
 
             // redirec admin/pneus
-            header('Location: /admininistration/membres');
+            header('Location: /administration/membres');
         }
     }
 }
